Supports authentication via /authenticate endpoint
The /authenticate endpoint is introduced and takes a query parameter
named 'token'. This token is used to receive information about the user
via the google environment.
The endpoint returnes the user in case he ist already known or creates a
new user if it is the first authentication attempt. The information
about the user is stored in the database.

Note: The API isn't secured at this point, this is only the first step
to validate users and provide a working authentication system.