Multiple tracks are supported and progresses are related to tracks
These tracks are identified by their internal id given by the mongo
database. This id can then be used to either retrieve them directly. In
addition to that, they have a status, whether they are active or not.

The tracks in its current version can not be modified, but this feature
will be supported later on.

Additionally: Connecting progresses with tracks by id

The tracks now can have progressesi attached to them. These progresses can either be added to the current track or to a track identified by the its ID. The progresses then can be retrieved by calling the GET method on either the track or a track identified by its ID.